converted AlertDialog to MD3
