+++ app/src/main/java/de/schweininchen/shintaikan/shintaikan/jetpack/NavigationDrawer.kt
@@ -6,26 +6,25 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.widget.Toast
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material.*
+import androidx.compose.material.Divider
+import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.*
+import androidx.compose.material.icons.outlined.Info
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
@@ -36,7 +35,7 @@ import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat.getSystemService
 import androidx.core.content.ContextCompat.startActivity
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
-import kotlinx.coroutines.launch
+import de.schweininchen.shintaikan.shintaikan.jetpack.components.navDrawer.DrawerItems
 
 val customPadding = 12.dp
 
@@ -59,7 +58,7 @@ fun drawerContent(
 
         if (openDialog.value) AboutAlertDialog { openDialog.value = false }
         if (openCustomDialog.value) CustomAlertDialog(
-            title = "",
+            title = "Weiteres",
             text = "Was Rüdiger noch sagen wollte:\nTiefer stehen, schneller schlagen! :)"
         ) {
             openCustomDialog.value = false
@@ -82,124 +81,10 @@ fun drawerContent(
                         .padding(top = 8.dp)
                 )
                 Divider(modifier = Modifier.padding(top = 4.dp, bottom = 4.dp))
-                Column {
-                    DrawerItem(
-                        "Start",
-                        Icons.Outlined.Info,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.HOME,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Trainingsplan",
-                        Icons.Outlined.DateRange,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.TRPLAN,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Gürtelprüfungen",
-                        Icons.Outlined.NorthEast,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.PRUEFUNGEN,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Ferientraining",
-                        Icons.Outlined.BeachAccess,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.FERIEN,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Nach den Sommerferien",
-                        Icons.Outlined.WbSunny,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.NACHSOFE,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Der Club / Wegbeschreibung",
-                        Icons.Outlined.Home,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.CLUBWEG,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Anfänger / Interressenten",
-                        Icons.Outlined.DirectionsWalk,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.ANFAENGER,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Vorführungen",
-                        Icons.Outlined.RemoveRedEye,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.VORFUEHRUNGEN,
-                        onClick = onClickMain,
-                    )
-                    DrawerItem(
-                        "Lehrgänge + Turniere",
-                        Icons.Outlined.People,
-                        selected = selectedMain,
-                        route = NavigationDrawerRoutes.LEHRGAENGE,
-                        onClick = onClickMain,
-                    )
-                    Divider()
-                    DrawerItem(
-                        "Infofilmchen",
-                        Icons.Outlined.Movie, externalLink = false, disabled = true
-                    ) { }
-                    DrawerItem(
-                        "Seefest 2019",
-                        Icons.Outlined.Movie, externalLink = false, disabled = true
-                    ) { }
-                    DrawerItem(
-                        "Mixfilm 2019",
-                        Icons.Outlined.Movie, externalLink = false, disabled = true
-                    ) { }
-                    Divider()
-                    DrawerItem(
-                        "Kontakt und Feedback",
-                        Icons.Outlined.Mail, externalLink = false, disabled = true
-                    ) { }
-                    DrawerItem(
-                        "Impressum",
-                        Icons.Outlined.AttachFile, true
-                    ) { linkToWebpage("https://shintaikan.de/?page_id=207", context = context) }
-                    DrawerItem(
-                        "Datenschutz",
-                        Icons.Outlined.Lock, true
-                    ) { linkToWebpage("https://shintaikan.de/?page_id=378", context = context) }
-                    DrawerItem(
-                        "Weiteres",
-                        Icons.Outlined.Face
-                    ) { openCustomDialog.value = true }
-                    //
-                    DrawerItem(
-                        "Über",
-                        Icons.Outlined.Info,
-                    ) {
-                        openDialog.value = true
-                    }
-                    Divider(modifier = Modifier.padding(bottom = 8.dp))
-                    Image(
-                        painter = painterResource(id = R.drawable.jetpackcompose_logo),
-                        contentDescription = "Jetpack Compose logo",
-                        modifier = Modifier
-                            .width(30.dp)
-                            .height(30.dp)
-                            .align(Alignment.CenterHorizontally)
-                            .clickable {
-                                showDebugInfo.value = true
-                                coroutineScope.launch {
-                                    listState.animateScrollToItem(index = listState.layoutInfo.totalItemsCount)
-                                }
-                            }
-                    )
-                    Box(modifier = Modifier.size(8.dp))
-                }
+                DrawerItems(
+                    selectedMain, onClickMain, context, openCustomDialog,
+                    openDialog, showDebugInfo, coroutineScope, listState
+                )
             }
             if (showDebugInfo.value) {
                 item {
@@ -209,6 +94,7 @@ fun drawerContent(
             }
         }
     }
+
 
 @Composable
 fun DebugInfo(vm: MyViewModel) {
@@ -240,81 +126,12 @@ fun DebugInfo(vm: MyViewModel) {
     }
 }
 
-
-@Composable
-private fun DrawerItem(
-    name: String,
-    icon: ImageVector,
-    externalLink: Boolean = false,
-    disabled: Boolean = false,
-    selected: NavigationDrawerRoutes? = null,
-    route: NavigationDrawerRoutes = NavigationDrawerRoutes.NONE,
-    onClick: (NavigationDrawerRoutes?) -> Unit
-) {
-    val color =
-        if (selected == route && !disabled) androidx.compose.material3.MaterialTheme.colorScheme.primary else Color(
-            0xFF898989
-        )
-    Row(
-        modifier = Modifier
-            .padding(top = 4.dp, bottom = 4.dp, start = 8.dp, end = 8.dp)
-            .clip(MaterialTheme.shapes.medium)
-            .clickable(onClick = { onClick(route) })
-            .background(
-                color = if (selected == route) androidx.compose.material3.MaterialTheme.colorScheme.primary.copy(
-                    alpha = .15f
-                ) else Color.Transparent
-            )
-            .padding(top = customPadding, bottom = customPadding, start = 12.dp, end = 12.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        if (disabled) {
-            Icon(imageVector = icon, "Drawer item", tint = Color(0xFFC9C9C9))
-            Text(
-                modifier = Modifier
-                    .padding(start = 20.dp)
-                    .weight(1f), text = name,
-                style = TextStyle(fontWeight = FontWeight.Bold, color = Color(0xFFC9C9C9))
-            )
-            if (externalLink) {
-                Box(contentAlignment = Alignment.CenterEnd) {
-                    Icon(
-                        imageVector = Icons.Outlined.ExitToApp,
-                        "",
-                        tint = Color(0xFFC9C9C9)
-                    )
-                }
-            }
-        } else {
-            Icon(imageVector = icon, "Drawer item", tint = color)
-            Text(
-                modifier = Modifier
-                    .padding(start = 20.dp)
-                    .weight(1f), text = name,
-                style = TextStyle(
-                    fontWeight = FontWeight.Bold,
-                    color = if (selected == route) androidx.compose.material3.MaterialTheme.colorScheme.primary else Color.Black
-                )
-            )
-            if (externalLink) {
-                Box(contentAlignment = Alignment.CenterEnd) {
-                    Icon(
-                        imageVector = Icons.Outlined.ExitToApp,
-                        "",
-                        tint = color
-                    )
-                }
-            }
-        }
-    }
-
-}
-
 @Composable
 fun CustomAlertDialog(title: String, text: String, onDissmiss: () -> Unit) {
-    AlertDialog(
+    androidx.compose.material3.AlertDialog(
         onDismissRequest = onDissmiss,
         title = { if (title.isNotEmpty()) Text(text = title) },
+        icon = { Icon(imageVector = Icons.Outlined.Info, contentDescription = "Info icon") },
         text = {
             Text(
                 text = text,
@@ -322,11 +139,11 @@ fun CustomAlertDialog(title: String, text: String, onDissmiss: () -> Unit) {
                 textAlign = TextAlign.Start
             )
         },
-        buttons = {
-            Button(
-                modifier = Modifier
+        confirmButton = {
+            androidx.compose.material3.TextButton(
+                /*modifier = Modifier
                     .fillMaxWidth()
-                    .padding(8.dp),
+                    .padding(8.dp),*/
                 onClick = onDissmiss
 
             ) {
@@ -341,7 +158,11 @@ fun CustomAlertDialog(title: String, text: String, onDissmiss: () -> Unit) {
 fun AboutAlertDialog(onDissmiss: () -> Unit) {
     val context = LocalContext.current
 
-    AlertDialog(
+    val gitHubButtonColors = ButtonDefaults.buttonColors(
+        containerColor = Color.DarkGray,
+    )
+
+    androidx.compose.material3.AlertDialog(
         onDismissRequest = onDissmiss,
         title = {
             Row {
@@ -361,14 +182,14 @@ fun AboutAlertDialog(onDissmiss: () -> Unit) {
         },
         text = {
             Column {
-                Button(
+                androidx.compose.material3.Button(
                     onClick = {
                         linkToWebpage(
                             "https://github.com/Mister-Muffin/Shintaikan",
                             context = context
                         )
                     },
-                    colors = ButtonDefaults.buttonColors(backgroundColor = Color.DarkGray),
+                    colors = gitHubButtonColors,
                     modifier = Modifier
                         .fillMaxWidth()
                         .align(Alignment.CenterHorizontally)
@@ -378,42 +199,27 @@ fun AboutAlertDialog(onDissmiss: () -> Unit) {
                 }
             }
         },
-        buttons = {
-            Row(
-                modifier = Modifier.padding(all = 8.dp),
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically,
+        dismissButton = {
+            androidx.compose.material3.OutlinedButton(
+                onClick = onDissmiss
             ) {
-                Button(
-                    modifier = Modifier
-                        .weight(1f)
-                        .padding(4.dp),
-                    border = BorderStroke(
-                        1.dp,
-                        androidx.compose.material3.MaterialTheme.colorScheme.primaryContainer
-                    ),
-                    colors = ButtonDefaults.outlinedButtonColors(contentColor = androidx.compose.material3.MaterialTheme.colorScheme.primaryContainer),
-                    onClick = {
-                        context.startActivity(
-                            Intent(
-                                context,
-                                OssLicensesMenuActivity::class.java
-                            )
+                Text("Close")
+            }
+
+        },
+        confirmButton = {
+            androidx.compose.material3.Button(
+                onClick = {
+                    context.startActivity(
+                        Intent(
+                            context,
+                            OssLicensesMenuActivity::class.java
                         )
-                    }
-
-                ) {
-                    Text("View licenses")
+                    )
                 }
-                Button(
-                    modifier = Modifier
-                        .weight(1f)
-                        .padding(4.dp),
-                    onClick = onDissmiss
 
-                ) {
-                    Text("Close")
-                }
+            ) {
+                Text("View licenses")
             }
         }
     )
@@ -421,9 +227,7 @@ fun AboutAlertDialog(onDissmiss: () -> Unit) {
 }
 
 fun linkToWebpage(uri: String, context: Context) {
-    //val context = ContextAmbient.current
     val openURL = Intent(Intent.ACTION_VIEW)
     openURL.data = Uri.parse(uri)
     startActivity(context, openURL, null)
-
 }
